+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## 0.1.2 (2021-08-31)
 
+### New Features
+- Adding the #ankify-root tag to a block makes it so all descendent blocks which are ankified will only include parents up to the block with the #ankify-root tag
+
+### Improvements
+- Added more syntax for defining cloze hints e.g. "Example of {clozed text[[::]]hint}", "Example of {clozed text[[::hint]]}", "Example of [[{]]clozed text[[::hint}]]"
+
 ### Fixes
 - Correctly parse more code block languages
+- Option tags closest to block now prioritized
 
 ## 0.1.1 (2021-05-23)